renaming code dir to src and fixing tests
