+++ tests/test_data_io.py
@@ -1,7 +1,7 @@
 import os
 import pandas as pd
 import pytest
-from utils.data_io import load_data, make_outdir
+from src.utils.data_io import load_data, make_outdir
 
 def test_load_data_csv(tmp_path):
     path = tmp_path / "test.csv"
